Putting server and LLM on different processes
