+++ vllm/commit_id.py
@@ -1,0 +1,1 @@
+__commit__ = "37d776606f7e434ac61758fd6601bb9df4ec505a"
